Move number, string and map decoding to separate modules
