+++ trace-utils/src/msgpack_decoder/mod.rs
@@ -1,4 +1,5 @@
 // Copyright 2024-Present Datadog, Inc. https://www.datadoghq.com/
 // SPDX-License-Identifier: Apache-2.0
 
+/// Decoding logic for V04 encoded trace payload
 pub mod v04;
