+++ trace-utils/src/msgpack_decoder/v04/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2024-Present Datadog, Inc. https://www.datadoghq.com/
 // SPDX-License-Identifier: Apache-2.0
 
+/// Decode a V04 encoded trace payload
 pub mod decoder;
+/// Error type used for decoding
 pub mod error;
-pub mod number;
